requirements.txt
streamlit>=1.28.0
google-generativeai>=0.3.0
pytesseract>=0.3.10
pdf2image>=1.16.3
PyPDF2>=3.0.1
python-docx>=1.1.0
nltk>=3.8.1
spacy>=3.7.0
langdetect>=1.0.9
plotly>=5.17.0
streamlit-authenticator>=0.2.3
pandas>=2.1.0
matplotlib>=3.8.0
Pillow>=10.0.0
